requirements.txt
schedule
beautifulsoup4
requests
requests_oauthlib
python-dotenv